refactor: invert provider contract dependency
