+++ src/tau_agent/harness.py
@@ -18,8 +18,8 @@ from tau_agent.messages import (
     ToolResultMessage,
     UserMessage,
 )
+from tau_agent.provider import ModelProvider
 from tau_agent.tools import AgentTool
-from tau_ai.provider import ModelProvider
 
 EventListener = Callable[[AgentEvent], Awaitable[None] | None]
 QueueMode = Literal["one_at_a_time", "all"]
