+++ src/tau_agent/provider.py
@@ -1,0 +1,32 @@
+"""Provider contract owned by Tau's portable agent layer."""
+
+from __future__ import annotations
+
+from collections.abc import AsyncIterator
+from typing import Protocol
+
+from tau_agent.messages import AgentMessage
+from tau_agent.provider_events import AssistantMessageEvent
+from tau_agent.tools import AgentTool
+
+
+class CancellationToken(Protocol):
+    def is_cancelled(self) -> bool:
+        """Return whether the current stream should stop."""
+        ...
+
+
+class ModelProvider(Protocol):
+    """Provider-neutral Pi-compatible model stream interface."""
+
+    def stream_response(
+        self,
+        *,
+        model: str,
+        system: str,
+        messages: list[AgentMessage],
+        tools: list[AgentTool],
+        signal: CancellationToken | None = None,
+    ) -> AsyncIterator[AssistantMessageEvent]:
+        """Stream one model response as assistant message events."""
+        ...
