+++ src/tau_ai/provider.py
@@ -1,32 +1,5 @@
-"""Provider protocol for Tau model adapters."""
+"""Public re-exports of the provider contract implemented by Tau adapters."""
 
-from __future__ import annotations
+from tau_agent.provider import CancellationToken, ModelProvider
 
-from collections.abc import AsyncIterator
-from typing import Protocol
-
-from tau_agent.messages import AgentMessage
-from tau_agent.tools import AgentTool
-from tau_ai.events import AssistantMessageEvent
-
-
-class CancellationToken(Protocol):
-    def is_cancelled(self) -> bool:
-        """Return whether the current stream should stop."""
-        ...
-
-
-class ModelProvider(Protocol):
-    """Provider-neutral Pi-compatible model stream interface."""
-
-    def stream_response(
-        self,
-        *,
-        model: str,
-        system: str,
-        messages: list[AgentMessage],
-        tools: list[AgentTool],
-        signal: CancellationToken | None = None,
-    ) -> AsyncIterator[AssistantMessageEvent]:
-        """Stream one model response as assistant message events."""
-        ...
+__all__ = ["CancellationToken", "ModelProvider"]
